clean files

+++ src/Http/Controllers/BlogController.php
@@ -86,12 +86,14 @@ class BlogController extends Controller
             'module_id'=>'sahak.avatar/membership',
             'slug' => "all_" . $blog->slug,
         ]);
-        $child=$frontPagesRepository->findBy('parent_id',$page->id);
         try{
+            if($page){
+            $child=$frontPagesRepository->findBy('parent_id',$page->id);
             Painter::findByVariation($child->template)->variations(false)->deleteVariation($child->template);
             Painter::findByVariation($page->template)->variations(false)->deleteVariation($page->template);
             $child->delete();
             $page->delete();
+            }
             $blog->delete();
             CreatePostsTable::down($blog->slug);
         }catch (\Exception $e){
